custom_addons_gitlab/requirements.txt
python-gitlab>=3.0